requirements/app.txt
Django==3.0.5
asgiref==3.2.7
Babel==2.8.0
backcall==0.1.0
beautifulsoup4==4.9.0
decorator==4.4.2
django-appconf==1.0.4
django-ckeditor==5.9.0
django-compressor==2.4
django-extensions==2.2.9
django-js-asset==1.2.2
django-phonenumber-field==4.0.0
django-simple-history==2.10.0
djangorestframework==3.11.0
drf-base64==2.0
drf-yasg==1.20.0
html5lib==1.0.1
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
parso==0.6.2
pexpect==4.8.0
phonenumbers==8.12.2
pickleshare==0.7.5
prompt-toolkit==3.0.5
psycopg2-binary==2.8.5
ptyprocess==0.6.0
pydot==1.4.1
Pygments==2.6.1
pyparsing==2.4.7
pytz==2019.3
rcssmin==1.0.6
rjsmin==1.1.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
traitlets==4.3.3
validate-email==1.3
wcwidth==0.1.9
webencodings==0.5.1

# No used requirements
# which will be used in future
djangorestframework-datatables==0.5.2

# debugging in Dev mode
django-debug-toolbar==2.2
